Enables support for distributing Bluebird as a binary library
This fixes #8 and adds a different solution for #6.

**Solution**

To allow frameworks to be used by multiple Swift version, we need to set:

```
BUILD_LIBRARY_FOR_DISTRIBUTION = YES
```

This isn't necessary for a Swift package (since they are build from source),
but it's useful for Carthage or other means of distributing binary frameworks.

The workaround is to use an `xcconfig` file and pass that as an argument to
`generate-xcodeproj`.

It was discovered that the `Bluebird_Info.plist` file is ovewritten when
running `generate-xcodeproj`. This means we cannot manually set the
`CFBundleVersion` there. The solution was to also add `CURRENT_PROJECT_VERSION`
to the `xcconfig` file.

**Testing**

I build `Bluebird.framework` in Xcode 11.3.1. Then in Xcode 11.4 I created a
test project and linked to the generated project. It built and linked
successfully.