Set APPLICATION_EXTENSION_API_ONLY to YES in Library.xcconfig (#10)
The Cocoapods podspec specifies `APPLICATION_EXTENSION_API_ONLY`, but the
Xcode project doesn't. This means that the resulting Carthage framework
isn't marked as extension safe and generates a warning when linked with
extension safe targets.

This change adds the setting to the `Library.xcconfig` file so that it
carries through when Carthage builds the dynamic framework.